server/job_controller/requirements.txt
flask
flask_restful
mysql.connector
configparser
google-api-python-client
google-cloud-speech
google-cloud-vision
requests
python-dotenv
opencv-python
moviepy
grpcio